requirements.txt
torch
transformers
gradio
pytesseract
opencv-python
Pillow
